+++ x/currency/codec.pb.go
@@ -6,6 +6,7 @@ package currency
 import (
 	fmt "fmt"
 	proto "github.com/gogo/protobuf/proto"
+	weave "github.com/iov-one/weave"
 	io "io"
 	math "math"
 )
@@ -24,7 +25,8 @@ const _ = proto.GoGoProtoPackageIsVersion2 // please upgrade the proto package
 // TokenInfo contains information about a single currency. It is used as an
 // alternative solution to hardcoding supported currencies information.
 type TokenInfo struct {
-	Name string `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
+	Metadata *weave.Metadata `protobuf:"bytes,1,opt,name=metadata,proto3" json:"metadata,omitempty"`
+	Name     string          `protobuf:"bytes,2,opt,name=name,proto3" json:"name,omitempty"`
 }
 
 func (m *TokenInfo) Reset()         { *m = TokenInfo{} }
@@ -60,6 +62,13 @@ func (m *TokenInfo) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_TokenInfo proto.InternalMessageInfo
 
+func (m *TokenInfo) GetMetadata() *weave.Metadata {
+	if m != nil {
+		return m.Metadata
+	}
+	return nil
+}
+
 func (m *TokenInfo) GetName() string {
 	if m != nil {
 		return m.Name
@@ -70,8 +79,9 @@ func (m *TokenInfo) GetName() string {
 // NewTokenInfoMsg will register a new currency. Ticker (currency symbol) can
 // be registered only once.
 type NewTokenInfoMsg struct {
-	Ticker string `protobuf:"bytes,1,opt,name=ticker,proto3" json:"ticker,omitempty"`
-	Name   string `protobuf:"bytes,2,opt,name=name,proto3" json:"name,omitempty"`
+	Metadata *weave.Metadata `protobuf:"bytes,1,opt,name=metadata,proto3" json:"metadata,omitempty"`
+	Ticker   string          `protobuf:"bytes,2,opt,name=ticker,proto3" json:"ticker,omitempty"`
+	Name     string          `protobuf:"bytes,3,opt,name=name,proto3" json:"name,omitempty"`
 }
 
 func (m *NewTokenInfoMsg) Reset()         { *m = NewTokenInfoMsg{} }
@@ -107,6 +117,13 @@ func (m *NewTokenInfoMsg) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_NewTokenInfoMsg proto.InternalMessageInfo
 
+func (m *NewTokenInfoMsg) GetMetadata() *weave.Metadata {
+	if m != nil {
+		return m.Metadata
+	}
+	return nil
+}
+
 func (m *NewTokenInfoMsg) GetTicker() string {
 	if m != nil {
 		return m.Ticker
@@ -129,17 +146,20 @@ func init() {
 func init() { proto.RegisterFile("x/currency/codec.proto", fileDescriptor_540c9a7fd55dd714) }
 
 var fileDescriptor_540c9a7fd55dd714 = []byte{
-	// 145 bytes of a gzipped FileDescriptorProto
+	// 205 bytes of a gzipped FileDescriptorProto
 	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xe2, 0x12, 0xab, 0xd0, 0x4f, 0x2e,
 	0x2d, 0x2a, 0x4a, 0xcd, 0x4b, 0xae, 0xd4, 0x4f, 0xce, 0x4f, 0x49, 0x4d, 0xd6, 0x2b, 0x28, 0xca,
-	0x2f, 0xc9, 0x17, 0xe2, 0x80, 0x89, 0x2a, 0xc9, 0x73, 0x71, 0x86, 0xe4, 0x67, 0xa7, 0xe6, 0x79,
-	0xe6, 0xa5, 0xe5, 0x0b, 0x09, 0x71, 0xb1, 0xe4, 0x25, 0xe6, 0xa6, 0x4a, 0x30, 0x2a, 0x30, 0x6a,
-	0x70, 0x06, 0x81, 0xd9, 0x4a, 0xb6, 0x5c, 0xfc, 0x7e, 0xa9, 0xe5, 0x70, 0x35, 0xbe, 0xc5, 0xe9,
-	0x42, 0x62, 0x5c, 0x6c, 0x25, 0x99, 0xc9, 0xd9, 0xa9, 0x45, 0x50, 0x85, 0x50, 0x1e, 0x5c, 0x3b,
-	0x13, 0x42, 0xbb, 0x93, 0xc4, 0x89, 0x47, 0x72, 0x8c, 0x17, 0x1e, 0xc9, 0x31, 0x3e, 0x78, 0x24,
-	0xc7, 0x38, 0xe1, 0xb1, 0x1c, 0xc3, 0x85, 0xc7, 0x72, 0x0c, 0x37, 0x1e, 0xcb, 0x31, 0x24, 0xb1,
-	0x81, 0x9d, 0x62, 0x0c, 0x08, 0x00, 0x00, 0xff, 0xff, 0x6f, 0x69, 0xf0, 0x93, 0xa4, 0x00, 0x00,
-	0x00,
+	0x2f, 0xc9, 0x17, 0xe2, 0x80, 0x89, 0x4a, 0xa9, 0xa4, 0x67, 0x96, 0x64, 0x94, 0x26, 0xe9, 0x25,
+	0xe7, 0xe7, 0xea, 0x67, 0xe6, 0x97, 0xe9, 0xe6, 0xe7, 0xa5, 0xea, 0x97, 0xa7, 0x26, 0x96, 0xa5,
+	0x22, 0xab, 0x57, 0xf2, 0xe1, 0xe2, 0x0c, 0xc9, 0xcf, 0x4e, 0xcd, 0xf3, 0xcc, 0x4b, 0xcb, 0x17,
+	0xd2, 0xe6, 0xe2, 0xc8, 0x4d, 0x2d, 0x49, 0x4c, 0x49, 0x2c, 0x49, 0x94, 0x60, 0x54, 0x60, 0xd4,
+	0xe0, 0x36, 0xe2, 0xd7, 0x03, 0x6b, 0xd1, 0xf3, 0x85, 0x0a, 0x07, 0xc1, 0x15, 0x08, 0x09, 0x71,
+	0xb1, 0xe4, 0x25, 0xe6, 0xa6, 0x4a, 0x30, 0x29, 0x30, 0x6a, 0x70, 0x06, 0x81, 0xd9, 0x4a, 0x59,
+	0x5c, 0xfc, 0x7e, 0xa9, 0xe5, 0x70, 0x03, 0x7d, 0x8b, 0xd3, 0x49, 0x33, 0x53, 0x8c, 0x8b, 0xad,
+	0x24, 0x33, 0x39, 0x3b, 0xb5, 0x08, 0x6a, 0x2a, 0x94, 0x07, 0xb7, 0x8b, 0x19, 0x61, 0x97, 0x93,
+	0xc4, 0x89, 0x47, 0x72, 0x8c, 0x17, 0x1e, 0xc9, 0x31, 0x3e, 0x78, 0x24, 0xc7, 0x38, 0xe1, 0xb1,
+	0x1c, 0xc3, 0x85, 0xc7, 0x72, 0x0c, 0x37, 0x1e, 0xcb, 0x31, 0x24, 0xb1, 0x81, 0xbd, 0x66, 0x0c,
+	0x08, 0x00, 0x00, 0xff, 0xff, 0x24, 0x00, 0x47, 0x84, 0x24, 0x01, 0x00, 0x00,
 }
 
 func (m *TokenInfo) Marshal() (dAtA []byte, err error) {
@@ -157,8 +177,18 @@ func (m *TokenInfo) MarshalTo(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.Name) > 0 {
+	if m.Metadata != nil {
 		dAtA[i] = 0xa
+		i++
+		i = encodeVarintCodec(dAtA, i, uint64(m.Metadata.Size()))
+		n1, err := m.Metadata.MarshalTo(dAtA[i:])
+		if err != nil {
+			return 0, err
+		}
+		i += n1
+	}
+	if len(m.Name) > 0 {
+		dAtA[i] = 0x12
 		i++
 		i = encodeVarintCodec(dAtA, i, uint64(len(m.Name)))
 		i += copy(dAtA[i:], m.Name)
@@ -181,14 +211,24 @@ func (m *NewTokenInfoMsg) MarshalTo(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.Ticker) > 0 {
+	if m.Metadata != nil {
 		dAtA[i] = 0xa
+		i++
+		i = encodeVarintCodec(dAtA, i, uint64(m.Metadata.Size()))
+		n2, err := m.Metadata.MarshalTo(dAtA[i:])
+		if err != nil {
+			return 0, err
+		}
+		i += n2
+	}
+	if len(m.Ticker) > 0 {
+		dAtA[i] = 0x12
 		i++
 		i = encodeVarintCodec(dAtA, i, uint64(len(m.Ticker)))
 		i += copy(dAtA[i:], m.Ticker)
 	}
 	if len(m.Name) > 0 {
-		dAtA[i] = 0x12
+		dAtA[i] = 0x1a
 		i++
 		i = encodeVarintCodec(dAtA, i, uint64(len(m.Name)))
 		i += copy(dAtA[i:], m.Name)
@@ -211,6 +251,10 @@ func (m *TokenInfo) Size() (n int) {
 	}
 	var l int
 	_ = l
+	if m.Metadata != nil {
+		l = m.Metadata.Size()
+		n += 1 + l + sovCodec(uint64(l))
+	}
 	l = len(m.Name)
 	if l > 0 {
 		n += 1 + l + sovCodec(uint64(l))
@@ -224,6 +268,10 @@ func (m *NewTokenInfoMsg) Size() (n int) {
 	}
 	var l int
 	_ = l
+	if m.Metadata != nil {
+		l = m.Metadata.Size()
+		n += 1 + l + sovCodec(uint64(l))
+	}
 	l = len(m.Ticker)
 	if l > 0 {
 		n += 1 + l + sovCodec(uint64(l))
@@ -278,6 +326,42 @@ func (m *TokenInfo) Unmarshal(dAtA []byte) error {
 		}
 		switch fieldNum {
 		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Metadata", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowCodec
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthCodec
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthCodec
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Metadata == nil {
+				m.Metadata = &weave.Metadata{}
+			}
+			if err := m.Metadata.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
 			if wireType != 2 {
 				return fmt.Errorf("proto: wrong wireType = %d for field Name", wireType)
 			}
@@ -364,6 +448,42 @@ func (m *NewTokenInfoMsg) Unmarshal(dAtA []byte) error {
 		switch fieldNum {
 		case 1:
 			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Metadata", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowCodec
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthCodec
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthCodec
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Metadata == nil {
+				m.Metadata = &weave.Metadata{}
+			}
+			if err := m.Metadata.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
 				return fmt.Errorf("proto: wrong wireType = %d for field Ticker", wireType)
 			}
 			var stringLen uint64
@@ -394,7 +514,7 @@ func (m *NewTokenInfoMsg) Unmarshal(dAtA []byte) error {
 			}
 			m.Ticker = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
-		case 2:
+		case 3:
 			if wireType != 2 {
 				return fmt.Errorf("proto: wrong wireType = %d for field Name", wireType)
 			}